+++ src/umpire/util/AllocationMap.cpp
@@ -9,6 +9,7 @@
 #include "umpire/util/FixedMallocPool.hpp"
 
 #include "umpire/util/Macros.hpp"
+#include "umpire/Replay.hpp"
 
 #include <sstream>
 #include <type_traits>
@@ -177,6 +178,7 @@ AllocationMap::insert(void* ptr, AllocationRecord record)
   std::lock_guard<std::mutex> lock(m_mutex);
 
   UMPIRE_LOG(Debug, "Inserting " << ptr);
+  UMPIRE_REPLAY("\"event\": \"allocation_map_insert\", \"payload\": { \"ptr\": \"" << ptr << "\", \"record_ptr\": \"" << record.ptr << "\", \"record_size\": \"" << record.size << "\", \"record_strategy\": \"" << record.strategy << "\" }");
 
   auto pair = m_map.insert(ptr, *this, record);
 
@@ -199,6 +201,7 @@ AllocationMap::find(void* ptr) const
   std::lock_guard<std::mutex> lock(m_mutex);
 
   UMPIRE_LOG(Debug, "Searching for " << ptr);
+  UMPIRE_REPLAY("\"event\": \"allocation_map_find\", \"payload\": { \"ptr\": \"" << ptr << "\" }");
 
   const AllocationRecord* alloc_record = doFindRecord(ptr);
 
@@ -273,6 +276,7 @@ AllocationMap::remove(void* ptr)
   AllocationRecord ret;
 
   UMPIRE_LOG(Debug, "Removing " << ptr);
+  UMPIRE_REPLAY("\"event\": \"allocation_map_remove\", \"payload\": { \"ptr\": \"" << ptr << "\" }");
 
   auto iter = m_map.find(ptr);
 
@@ -303,6 +307,7 @@ AllocationMap::clear()
   std::lock_guard<std::mutex> lock(m_mutex);
 
   UMPIRE_LOG(Debug, "Clearing");
+  UMPIRE_REPLAY("\"event\": \"allocation_map_clear\"");
 
   m_map.clear();
   m_size = 0;