+++ CHANGELOG.md
@@ -20,6 +20,8 @@ Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to
 
 ### Added
 
+- Added ability to replay allocation maps for testing purposes.
+
 - CI builds for Mac, Linux and Windows via Azure Pipelines
 
 - HCC stage in Docker file.